moved parcelTable to child of map; added clear selection button for selected parcel. #31.
