+++ frontend/src/index.css
@@ -13,6 +13,7 @@ html, body, #root, .App {
 }
 
 .map-container {
+  position: absolute;
   /*flex-grow: 1;*/
   height: 100%;
   width: 60%;
@@ -54,6 +55,20 @@ html, body, #root, .App {
   padding: 0.5rem;
   margin: 0;
   filter: drop-shadow(2px 2px 2px grey);
+}
+
+.layer-info {
+  background-color: rgba(255,255,255,0.9);
+  position: absolute;
+  top: 1rem;
+  right: 1rem;
+  border-radius: 5px;
+}
+
+.layer-info p {
+  padding: 0.5rem;
+  margin: 0;
+  font-weight: 600;
 }
 
 /*.edit-wallpaper {
